src/requirements.txt
black==24.3.0
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
fonttools==4.51.0
gevent==24.2.1
geventhttpclient==2.0.12
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
locust==2.24.1
MarkupSafe==2.1.5
matplotlib==3.8.4
msgpack==1.0.8
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
paho-mqtt==2.0.0
pandas==2.2.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
zope.event==5.0
zope.interface==6.2
